debug

+++ dl_helper/tests/rl/clear_run_all.py
@@ -1,11 +1,13 @@
 import os
+import sys
 import shutil
 import subprocess 
 
-def run():
-    # 清理 /root/alist_data/rl_learning_process 目录下的所有文件
-    shutil.rmtree('/root/alist_data/rl_learning_process')
-    os.makedirs('/root/alist_data/rl_learning_process')
+def run(need_clear):
+    if need_clear:
+        # 清理 /root/alist_data/rl_learning_process 目录下的所有文件
+        shutil.rmtree('/root/alist_data/rl_learning_process')
+        os.makedirs('/root/alist_data/rl_learning_process')
     
     # 训练服务
     for py_file in ['20241203_dddqn.py', '20241216_per_dqn.py', '20241217_c51.py']:
@@ -17,4 +19,10 @@ def run():
     subprocess.run(f'python3 /root/code/dl_helper/dl_helper/rl/net_center.py', check=True)
 
 if __name__ == '__main__':
-    run()
+    need_clear = False
+
+    # 获取命令行参数
+    if len(sys.argv) > 1 and sys.argv[1] == 'clear':
+        need_clear = True
+    
+    run(need_clear)
